folding_rl/pyproject.toml
[project]
name = "folding-rl"
version = "0.1.0"
description = "Protein folding via reinforcement learning on a 3D voxel grid"
readme = "README.md"
authors = [
    { name = "Finn Lueth", email = "finn@lueth.net" }
]
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1",
    "gymnasium>=0.29",
    "biopython>=1.83",
    "numpy>=1.26",
    "wandb>=0.16",
    "einops>=0.7",
    "scipy>=1.12",
    "pytorch-lightning>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/folding_rl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
